stat.rs: implement `Index<Stat> for [usize]`

+++ src/clause.rs
@@ -433,7 +433,7 @@ impl ClauseDBIF for ClauseDB {
                 c.kill(touched);
             }
         }
-        state.stats[Stat::Reduction as usize] += 1;
+        state.stats[Stat::Reduction] += 1;
         self.garbage_collect();
     }
     fn simplify(
@@ -464,9 +464,9 @@ impl ClauseDBIF for ClauseDB {
             }
         }
         self.garbage_collect();
-        state.stats[Stat::SatClauseElimination as usize] += 1;
+        state.stats[Stat::SatClauseElimination] += 1;
         if elim.is_running() {
-            state.stats[Stat::ExhaustiveElimination as usize] += 1;
+            state.stats[Stat::ExhaustiveElimination] += 1;
             self.reset_lbd(vars, &mut state.lbd_temp);
             elim.stop(self, vars);
         }
